Added support for gopass as a backend

+++ autovault.zsh
@@ -4,6 +4,14 @@ CWD="$(cd "$(dirname "${(%):-%x}")";pwd -P)"
 SCRIPT="$(readlink -f ${(%):-%x})"
 TMP_BASE="${TMPDIR:-/tmp/}$(basename ${SCRIPT%.*} 2>/dev/null)"
 
+# Backend config + defaults
+BACKEND="${AUTOVAULT_BACKEND:-1pass}"
+BACKEND_PREFIX="${AUTOVAULT_BACKEND_PREFIX}"
+case "${BACKEND}" in
+    1pass  ) BACKEND_PREFIX="${BACKEND_PREFIX:-cli:vaulted:}" ;;
+    gopass ) BACKEND_PREFIX="${BACKEND_PREFIX:-${GOPASS_MOUNT:-inst/}cli/vaulted/}" ;;
+esac
+
 
 # =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-= #
 # Super ximple XOR encryption for our temp files  #
@@ -29,13 +37,77 @@ dec() {
 }
 
 # =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-= #
+# Backends                                        #
+# =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-= #
+
+backend_1pass() {
+    local action="${1}"
+    local vault="${2}"
+    local entry="${3}"
+    local value="${4}"
+    case "${1}" in
+        fetch )
+            [[ "${entry}" == "password" ]] && echo $(1pass -p "${BACKEND_PREFIX}${vault}" password) && return 0
+            [[ "${entry}" == "otp" ]]      && echo $(1pass -p "${BACKEND_PREFIX}${vault}" totp)     && return 0
+            ;;
+        write )
+            if [[ "${entry}" == "password" ]]; then
+                local opsession="$(gpg -qdr ${$(grep self_key ${HOME}/.1pass/config)##*=} ~/.1pass/cache/_session.gpg)"
+                if [[ -z "$(backend_${BACKEND} fetch "${vault}" password)" ]]; then
+                    echo "Creating an entry named '${BACKEND_PREFIX}${vault}' in 1Password. Be sure to update it with MFA if needed" > /dev/tty
+                    echo "${opsession}" | \
+                        op create item Password "$(op encode <<< '{"password":"'"${value}"'"}')" --title="${BACKEND_PREFIX}${vault}" 2>&1 > /dev/null
+                    echo -en "Refreshing 1pass cache..." > /dev/tty; 1pass -r > /dev/null; echo "done" > /dev/tty
+                else
+                    echo "Updating existing entries with the '${BACKEND}' backend is not currently supported (see: https://discussions.agilebits.com/discussion/84324/cli-delete-update-logins)" > /dev/tty
+                    return 1
+                fi
+            else
+                echo "The specified action '${action}' is not currently supported for writing with the '${BACKEND}' backend" > /dev/tty
+                return 1
+            fi
+            ;;
+    esac
+}
+
+backend_gopass() {
+    local action="${1}"
+    local vault="${2}"
+    local entry="${3}"
+    local value="${4}"
+    case "${1}" in
+        fetch )
+            [[ "${entry}" == "password" ]] && echo $(gopass show -o "${BACKEND_PREFIX}${vault}" 2>/dev/null)                 && return 0
+            [[ "${entry}" == "otp" ]]      && echo $(gopass otp     "${BACKEND_PREFIX}${vault}" 2>/dev/null | cut -f1 -d' ') && return 0
+            ;;
+        write )
+            # Write a password
+            if [[ "${entry}" == "password" ]]; then
+                if [[ -z "$(backend_${BACKEND} fetch "${vault}" password)" ]]; then
+                    echo "Creating an entry named '${BACKEND_PREFIX}${vault}' in gopass. Be sure to update it with MFA if needed" > /dev/tty
+                    gopass insert -f "${BACKEND_PREFIX}${vault}" <<< "${value}"
+                else
+                    echo "Updating existing entries with the '${BACKEND}' backend is not yet supported" > /dev/tty
+                    return 1
+                fi
+
+            # Write a TOTP
+            else
+                echo "The specified action '${action}' is not currently supported for writing with the '${BACKEND}' backend" > /dev/tty
+                return 1
+            fi
+            ;;
+    esac
+}
+
+# =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-= #
 # The askpass responder                           #
 # =-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-= #
 
 ask() {
     if [[ ${1} =~ "'([^']+)' ((new|confirm) password|password|MFA token)([^:]+)?:" ]]; then
         local vault="${match[1]}"
-        local pwhsh="$(shasum <<< "${TMP_BASE}.$(stat -c %Y ${SCRIPT})")"
+        local pwhsh="$(shasum <<< "${TMP_BASE}.$(stat -c %Y ${SCRIPT}).${vault}")"
         local pwtmp="${TMP_BASE}.$(awk '{print substr($1,1,12)}' <<< ${pwhsh})"
         local pwkey="$(awk '{print substr($1,length($1)-11,12)}' <<< ${pwhsh})"
 
@@ -58,17 +130,20 @@ ask() {
                     fi
                 done
             elif [[ "${match[3]}" == "confirm" ]]; then
-                local pw="$(cat ${pwtmp} | dec ${pwkey})"
-                rm ${pwtmp}
-                echo -n "${pw}"
+                # Fetch the validated password
+                local pw=""
+                if [[ -f "${pwtmp}" ]]; then
+                    pw="$(cat ${pwtmp} | dec ${pwkey})"
+                    rm ${pwtmp}
+                else
+                    echo "Something went wrong obtaining the entered password...falling back to generating a random password for you" > /dev/tty
+                    pw="$(head -c32 /dev/urandom | shasum -a 256 | cut -f1 -d' ')"
+                fi
 
-                # Save an entry in 1password
-                local item_pass="$(1pass -p "cli:vaulted:${vault}" password)"
-                local opsession="$(gpg -qdr ${$(grep self_key ${HOME}/.1pass/config)##*=} ~/.1pass/cache/_session.gpg)"
-                if [[ -z "${item_pass}" ]]; then
-                    echo "Creating an entry named 'cli:vaulted:${vault}' in 1Password. Be sure to update it with MFA if needed" > /dev/tty
-                    echo "${opsession}" | op create item Password "$(echo '{"password":"test"}' | op encode)" --title="cli:vaulted:${vault}" 2>&1 > /dev/null
-                    echo -en "Refreshing 1pass cache..." > /dev/tty; 1pass -r > /dev/null; echo "done" > /dev/tty
+                # Save the password using the chosen backend
+                if [[ -n "${pw}" ]]; then
+                    backend_${BACKEND} write "${vault}" password "${pw}"
+                    echo -n "${pw}"
                 fi
             fi
         fi
@@ -76,8 +151,8 @@ ask() {
         #
         # Inject requested credential from 1password
         case "${match[2]}" in
-            password ) echo $(1pass -p "cli:vaulted:${vault}" password) ;;
-            *token   ) echo $(1pass -p "cli:vaulted:${vault}" totp)     ;;
+            password ) backend_${BACKEND} fetch "${vault}" password ;;
+            *token   ) backend_${BACKEND} fetch "${vault}" otp      ;;
         esac
     else
         echo "No match found for: ${1}" >> ${CWD}/.autovault.log
